+++ services/dynamic-sidecar/src/simcore_service_dynamic_sidecar/core/settings.py
@@ -4,10 +4,8 @@ from functools import lru_cache
 from pathlib import Path
 from typing import cast
 
-from common_library.pydantic_settings_validators import (
-    validate_timedelta_in_legacy_mode,
-)
 from common_library.pydantic_type_adapters import ByteSizeAdapter
+from common_library.pydantic_validators import validate_timedelta_in_legacy_mode
 from models_library.basic_types import BootModeEnum, PortInt
 from models_library.callbacks_mapping import CallbacksMapping
 from models_library.products import ProductName
